--- controllers/uffizzicluster/ingress.go
@@ -1,1 +1,0 @@
-package uffizzicluster
